rename scrollbar component
