+++ design/ui/styles/scrollbar/scrollbar.module.scss
